finished Sythisizer Application

+++ CS6011/Week2/Day6/Synthesizer/src/main/java/com/example/synthesizer/LinearRamp.java
@@ -5,7 +5,7 @@ public class LinearRamp implements AudioComponent{
     float start_;
     float stop_;
 
-    LinearRamp( int start, int stop){
+    LinearRamp( float start, float stop){
         start_ = start;
         stop_ = stop;
     }
